all goroutines receive a context and a wait group. - Refactor the http server to support context. - RDD http client now supports context

+++ pkg/monitoring/mapping.go
@@ -96,7 +96,7 @@ func MakeConfigSetMapping(
 					"observation_payment": int32(state.Config.Billing.ObservationPayment),
 				},
 				"validator":          state.Config.Validator[:],
-				"flagging_threshold": int(state.Config.FlaggingThreshold),
+				"flagging_threshold": int64(state.Config.FlaggingThreshold),
 			},
 			"oracles":              formatOracles(state.Oracles),
 			"leftover_payment":     formatLeftovers(state.LeftoverPayment, state.LeftoverPaymentLen),
